==> docs/notebooks/1-exploring-data.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt

df = pd.read_csv('https://www.openml.org/data/get_csv/16826755/phpMYEkMl')
df.shape

df.head()

df.replace('?', np.nan, inplace=True)

df.dtypes

Observe that age and fare are object types due to nan values

df['age'] = df['age'].astype('float')
df['fare'] = df['fare'].astype('float')

target = 'survived'
num_vars = [col for col in df.columns if df[col].dtype != object and col != target]
cat_vars = [col for col in df.columns if df[col].dtype == object]

len(num_vars) + len(cat_vars) + 1 == df.shape[1]

for col in df.columns:
    print(col)
    print('='*len(col))
    print(df[col].value_counts(dropna=False))
    print()

df.isna().sum()/df.shape[0]

def extract_title(input_string):
    return input_string.split(',')[1].split('.')[0].strip()

df['title'] = df['name'].apply(extract_title)

df['title'].value_counts()

df['cabin']
